Add SearchBox as it's own component
