+++ api/web/src/components/CloudTAK/Map.vue
@@ -104,7 +104,7 @@
                         stroke='1'
                         style='margin: 5px 8px'
                         class='cursor-pointer hover-button'
-                        @click='searchBox.shown = !searchBox.shown'
+                        @click='searchBoxShown = !searchBoxShown'
                     />
 
                     <div
@@ -198,31 +198,16 @@
                 @close='pointInput = false'
             />
 
-            <div
-                v-if='searchBox.shown'
-                class='position-absolute text-white bg-dark rounded'
+            <SearchBox
+                v-if='searchBoxShown'
                 style='
                     z-index: 1;
                     top: 8px;
                     left: 70px;
                     width: 200px;
                 '
-            >
-                <TablerInput
-                    v-model='searchBox.query.filter'
-                    class='mt-0'
-                    placeholder='Place Search'
-                    icon='search'
-                />
+            />
 
-                <div
-                    v-for='item of searchBox.results'
-                    :key='item.magicKey'
-                    class='col-12 px-2 py-2 hover-button cursor-pointer'
-                    @click='fetchSearch(item.text, item.magicKey)'
-                    v-text='item.text'
-                />
-            </div>
 
             <div
                 v-if='drawMode === "point"'
@@ -479,6 +464,7 @@ import {ref, watch, computed, onMounted, onBeforeUnmount, useTemplateRef } from 
 import {useRoute, useRouter } from 'vue-router';
 import CoTVideo from './util/Video.vue';
 import WarnChannels from './util/WarnChannels.vue';
+import SearchBox from './util/SearchBox.vue';
 import WarnConfiguration from './util/WarnConfiguration.vue';
 import Status from '../util/Status.vue';
 import CoordInput from './CoordInput.vue';
@@ -549,22 +535,7 @@ const warnChannels = ref<boolean>(false)
 // Show a popup if role/groups hasn't been set
 const warnConfiguration = ref<boolean>(false);
 
-const searchBox = ref<{
-    shown: boolean,
-    query: {
-        filter: string,
-    },
-    results: Array<{
-        text: string
-        magicKey: string
-    }>
-}>({
-    shown: false,
-    query: {
-        filter: '',
-    },
-    results: []
-});
+const searchBoxShown = ref(false);
 const drawMode = ref<string>('static') // Set the terra-draw mode to avoid getMode() calls
 const drawModePoint = ref<string>('u-d-p');
 const pointInput = ref<boolean>(false);
@@ -630,10 +601,6 @@ watch(mapStore.radial, () => {
         locked.value.pop();
     }
 })
-
-watch(searchBox.value.query, async () => {
-    await fetchSearch();
-});
 
 onMounted(async () => {
     // ensure uncaught errors in the stack are captured into vue context
@@ -803,33 +770,6 @@ function fileUpload(event: string) {
     upload.value.shown = false;
     const imp = JSON.parse(event) as { id: string };
     router.push(`/menu/imports/${imp.id}`)
-}
-
-async function fetchSearch(query?: string, magicKey?: string) {
-    if (!magicKey || !query) {
-        const url = stdurl('/api/search/suggest');
-        url.searchParams.append('query', searchBox.value.query.filter);
-        searchBox.value.results = ((await std(url)) as SearchSuggest).items;
-    } else {
-        const url = stdurl('/api/search/forward');
-        url.searchParams.append('query', query);
-        url.searchParams.append('magicKey', magicKey);
-        const items = ((await std(url)) as SearchForward).items;
-
-        searchBox.value.shown = false;
-        searchBox.value.query.filter = '';
-        searchBox.value.results = [];
-
-        if (items.length) {
-            mapStore.map.fitBounds([
-                [items[0].extent.xmin, items[0].extent.ymin],
-                [items[0].extent.xmax, items[0].extent.ymax],
-            ], {
-                duration: 0,
-                padding: {top: 25, bottom:25, left: 25, right: 25}
-            });
-        }
-    }
 }
 
 function getLocation() {
